# backend/openapi.yaml
...
openapi: 3.0.3
info:
  title: EduControl API
  version: '1.0'
  description: |
    Gestión de cursos, estudiantes y profesores.

    **Autorización.** Casi todo pide `Authorization: Bearer <token>`. El rol se
    lee de la base de datos en cada petición, no del claim del token: a quien
    degradan pierde el acceso de inmediato en vez de conservarlo hasta que
    caduque.

    **Propiedad.** El rol dice qué clase de usuario entra; de quién es el
    recurso lo decide el controlador leyéndolo. Un profesor solo toca sus
    cursos; un estudiante, sus matrículas.

    **Paginación.** Los listados devuelven `total`, `pagina`, `limite` y
    `paginas`, con un tope duro de 100 por página.
  license:
    name: MIT
    url: https://github.com/JohanParadaC/educontrol/blob/main/LICENSE
servers:
  - url: /
    description: El mismo origen que sirve la aplicación
...
security:
  - bearerAuth: []
paths:
...
  /api/usuarios:
    post:
      operationId: registrarUsuario
      tags: [Usuarios]
      summary: Registro público
      description: >
        Nunca crea un `admin`, venga como venga el cuerpo. `profesor` sí, pero
        exige la clave del centro (`profesorClave`).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, correo, contraseña]
              properties:
                nombre: { type: string }
                correo: { type: string, format: email }
                contraseña: { type: string, minLength: 6 }
                rol: { type: string, enum: [estudiante, profesor] }
                profesorClave: { type: string }
      responses:
        '201':
          description: Cuenta creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  usuario: { $ref: '#/components/schemas/Usuario' }
        '400':
          description: Datos inválidos o correo ya registrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Clave de profesor inválida
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      operationId: listarUsuarios
      tags: [Usuarios]
      summary: Listar usuarios
      parameters:
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/Limite'
        - name: rol
          in: query
          schema: { type: string, enum: [estudiante, profesor, admin] }
      responses:
        '200':
          description: Página de usuarios
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Paginacion'
                  - type: object
                    properties:
                      usuarios:
                        type: array
                        items: { $ref: '#/components/schemas/Usuario' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/Prohibido' }
...
  /api/inscripciones/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: obtenerInscripcion
      tags: [Inscripciones]
      summary: Una matrícula por id
      description: >
        Con el mismo alcance que el listado. Quien no puede verla recibe 404 y
        no 403: tampoco tiene por qué saber que existe.
      responses:
        '200':
          description: La matrícula
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  inscripcion: { $ref: '#/components/schemas/Inscripcion' }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
    delete:
      operationId: borrarInscripcion
      tags: [Inscripciones]
      summary: Dar de baja una matrícula
      description: >
        El dueño de la matrícula o un administrador. El profesor todavía no:
        echar a un alumno de clase es otra cosa y no está decidida.
      responses:
        '200':
          description: Baja hecha
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Mensaje' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        `Authorization: Bearer <token>`. Es la única cabecera de sesión que se
        acepta: la antigua `x-token` se retiró, porque dos puertas para lo
        mismo son el doble de superficie.
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: ObjectId de 24 caracteres hexadecimales
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
    Pagina:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limite:
      name: limit
      in: query
      description: Tope duro de 100; pedir más se recorta en silencio
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
  responses:
    NoAutenticado:
      description: Falta el token o no es válido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Prohibido:
      description: El rol o la propiedad del recurso no lo permiten
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NoEncontrado:
      description: No existe
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    DatosInvalidos:
      description: >
        Validación fallida. `errors` lleva un campo por cada problema, tal y
        como lo devuelve express-validator.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorValidacion' }
  schemas:
    Error:
      type: object
      properties:
        ok: { type: boolean, example: false }
        msg: { type: string }
    ErrorValidacion:
      type: object
      properties:
        ok: { type: boolean, example: false }
        errors:
          type: object
          additionalProperties:
            type: object
            properties:
              type: { type: string }
              msg: { type: string }
              path: { type: string }
              location: { type: string }
    Mensaje:
      type: object
      properties:
        ok: { type: boolean, example: true }
        msg: { type: string }
    Paginacion:
      type: object
      properties:
        ok: { type: boolean, example: true }
        total: { type: integer }
        pagina: { type: integer }
        limite: { type: integer }
        paginas: { type: integer }
...
    Usuario:
      type: object
      description: >
        El modelo renombra `_id` a `id` al serializar y nunca incluye la
        contraseña.
      properties:
        id: { type: string }
        nombre: { type: string }
        correo: { type: string, format: email }
        rol: { type: string, enum: [estudiante, profesor, admin] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Curso:
      type: object
      description: >
        El backend llama `nombre` a lo que la interfaz llama `titulo`; la
        traducción vive solo en el frontend.
      properties:
        _id: { type: string }
        nombre: { type: string, maxLength: 120 }
        descripcion: { type: string, maxLength: 500 }
        profesor:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Usuario'
        cupoMaximo:
          type: integer
          minimum: 1
          description: Ausente es «sin límite» — nunca 0, que sería «no cabe nadie»
        estado: { type: string, enum: [abierto, cerrado, archivado], default: abierto }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Inscripcion:
      type: object
      properties:
        _id: { type: string }
        estudiante:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Usuario'
        curso:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Curso'
        fecha:
          type: string
          format: date-time
          description: Se conserva por compatibilidad; para código nuevo, `createdAt`
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...